requirements.txt
bleach==2.1.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-b2storage==1.0
gunicorn==19.8.1
html5lib==1.0.1
Markdown==2.6.11
Pillow==5.1.0
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.4
six==1.11.0
static3==0.7.0
Unipath==1.1
webencodings==0.5.1
whitenoise==3.3.1
